fix(models): add missing Integer import to core and tenants models
Fix ruff F821 errors (undefined name 'Integer') in:
- app/core/models.py: Add Integer to sqlalchemy imports
- app/modules/tenants/models.py: Add Integer and remove unused Numeric import

+++ app/modules/tenants/models.py
@@ -10,7 +10,7 @@ import enum
 from datetime import datetime
 from sqlalchemy import (
     Column, String, Text, Boolean, DateTime,
-    Enum, Float, Numeric
+    Enum, Float, Integer
 )
 from app.core.database import Base
 from app.core.types import UniversalUUID, JSON
